requirements.txt

streamlit
python-docx
PyMuPDF
spacy
language-tool-python
scikit-learn
sentence-transformers
pdfminer.six
